backend/requirements.txt
dnspython==2.8.0
Flask==2.3.3
flask-cors==6.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
gunicorn==23.0.0
flask-pymongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
PyJWT==2.10.1
pymongo==4.15.4
python-dotenv==1.2.1
Werkzeug==3.1.3
sendgrid==6.11.0
python-http-client==3.3.7